Add media parameter to Playlist.Item directory type
Allow complete control from remote of the media to add, now permit a correct Player.Open on directories with mixed items.
Also permit to have metadata loaded when files are added from Player.Open
The cases where user force a wrong type for a specific playlist type is already handled in the code and return the FailedToExecute error.